Move steps to a separate file and refactor
